+++ Sources/OpenKey/macOS/ModernKey/OpenKeyManager.m
@@ -25,6 +25,7 @@ extern NSString* ConvertUtil(NSString* str);
 
 }
 static BOOL _isInited = NO;
+static BOOL _engineInited = NO;
 
 static CFMachPortRef      eventTap;
 static CGEventMask        eventMask;
@@ -35,11 +36,23 @@ static CFRunLoopSourceRef runLoopSource;
 }
 
 +(BOOL)initEventTap {
-    if (_isInited)
-        return true;
+    if (_isInited) {
+        if (eventTap && CFMachPortIsValid(eventTap)) {
+            if (!CGEventTapIsEnabled(eventTap)) {
+                CGEventTapEnable(eventTap, true);
+            }
+            if (CGEventTapIsEnabled(eventTap)) {
+                return YES;
+            }
+        }
+        [self stopEventTap];
+    }
     
     //init modernKey
-    OpenKeyInit();
+    if (!_engineInited) {
+        OpenKeyInit();
+        _engineInited = YES;
+    }
     
     // Create an event tap. We are interested in key presses.
     eventMask = ((1 << kCGEventKeyDown) |
@@ -55,7 +68,7 @@ static CFRunLoopSourceRef runLoopSource;
                                 0,
                                 eventMask,
                                 OpenKeyCallback,
-                                NULL);
+                                &eventTap);
     
     if (!eventTap) {
         
@@ -63,29 +76,48 @@ static CFRunLoopSourceRef runLoopSource;
         return NO;
     }
     
-    _isInited = YES;
-    
     // Create a run loop source.
     runLoopSource = CFMachPortCreateRunLoopSource(kCFAllocatorDefault, eventTap, 0);
+    if (!runLoopSource) {
+        CFMachPortInvalidate(eventTap);
+        CFRelease(eventTap);
+        eventTap = nil;
+        return NO;
+    }
     
     // Add to the current run loop.
     CFRunLoopAddSource(CFRunLoopGetCurrent(), runLoopSource, kCFRunLoopCommonModes);
     
     // Enable the event tap.
     CGEventTapEnable(eventTap, true);
+    if (!CGEventTapIsEnabled(eventTap)) {
+        CFRunLoopRemoveSource(CFRunLoopGetCurrent(), runLoopSource, kCFRunLoopCommonModes);
+        CFRelease(runLoopSource);
+        runLoopSource = nil;
+        CFMachPortInvalidate(eventTap);
+        CFRelease(eventTap);
+        eventTap = nil;
+        return NO;
+    }
+
+    _isInited = YES;
     
     return YES;
 }
 
 +(BOOL)stopEventTap {
     if (_isInited) { //release all object
-        CFRunLoopRemoveSource(CFRunLoopGetCurrent(), runLoopSource, kCFRunLoopCommonModes);
-        CFRelease(runLoopSource);
-        runLoopSource = nil;
+        if (runLoopSource) {
+            CFRunLoopRemoveSource(CFRunLoopGetCurrent(), runLoopSource, kCFRunLoopCommonModes);
+            CFRelease(runLoopSource);
+            runLoopSource = nil;
+        }
         
-        CFMachPortInvalidate(eventTap);
-        CFRelease(eventTap);
-        eventTap = nil;
+        if (eventTap) {
+            CFMachPortInvalidate(eventTap);
+            CFRelease(eventTap);
+            eventTap = nil;
+        }
         
         _isInited = false;
     }